Moved bitmap functions to util and removed a lot of duplicate code
