+++ src/d3d12/d3d12_heap.cpp
@@ -1,42 +1,10 @@
 #include "d3d12_functions.hpp"
 
 #include "d3d12_defines.hpp"
+#include "../util/bitmap_allocator.hpp"
 
 namespace wr::d3d12
 {
-	namespace internal 
-	{
-		inline std::uint64_t IndexFromBit(std::uint64_t frame) 
-		{
-			return frame / (8 * 8);
-		}
-
-		inline std::uint64_t OffsetFromBit(std::uint64_t frame) 
-		{
-			return frame % (8 * 8);
-		}
-
-		void SetPage(std::vector<uint64_t>* bitmap, std::uint64_t frame) 
-		{
-			std::uint64_t idx = IndexFromBit(frame);
-			std::uint64_t off = OffsetFromBit(frame);
-			bitmap->operator[](idx) |= (1Ui64 << off);
-		}
-
-		void ClearPage(std::vector<uint64_t>* bitmap, std::uint64_t frame) 
-		{
-			std::uint64_t idx = IndexFromBit(frame);
-			std::uint64_t off = OffsetFromBit(frame);
-			bitmap->operator[](idx) &= ~(1Ui64 << off);
-		}
-
-		bool TestPage(std::vector<uint64_t>* bitmap, std::uint64_t frame) 
-		{
-			std::uint64_t idx = IndexFromBit(frame);
-			std::uint64_t off = OffsetFromBit(frame);
-			return (bitmap->operator[](idx) & (1Ui64 << off));
-		}
-	}
 
 	Heap<HeapOptimization::SMALL_BUFFERS>* CreateHeap_SBO(Device* device, std::uint64_t size_in_bytes, ResourceType resource_type, unsigned int versioning_count)
 	{
@@ -241,67 +209,9 @@ namespace wr::d3d12
 		auto frame_count = heap->m_heap_size / heap->m_alignment;
 		auto needed_frames = aligned_size / heap->m_alignment*heap->m_versioning_count;
 
-		std::uint64_t start_frame = 0;
-		bool counting = false;
-		bool found = false;
-		int free_frames = 0;
+		auto start_frame = util::FindFreePage(heap->m_bitmap, frame_count, needed_frames);
 
-		for (std::uint64_t i = 0; i <= internal::IndexFromBit(frame_count); ++i) 
-		{
-			if (heap->m_bitmap[i] != 0Ui64) 
-			{
-				for (std::uint64_t j = 0; j < 64; ++j) 
-				{
-					std::uint64_t to_test = 1Ui64 << j;
-					if (i * 64 + j >= frame_count)
-						break;
-
-					if ((heap->m_bitmap[i] & to_test))
-					{
-						if (counting) 
-						{
-							free_frames++;
-							if (free_frames == needed_frames) 
-							{
-								found = true;
-								break;
-							}
-						}
-						else
-						{
-							if (needed_frames == 1) 
-							{
-								found = true;
-								start_frame = i * 64 + j;
-								break;
-							}
-							else 
-							{
-								start_frame = i * 64 + j;
-								counting = true;
-								free_frames = 1;
-							}
-						}
-					}
-					else 
-					{
-						counting = 0;
-						free_frames = 0;
-					}
-				}
-			}
-			else 
-			{
-				counting = false;
-				free_frames = 0;
-			}
-			if (found == true)
-			{
-				break;
-			}
-		}
-
-		if (found == false)
+		if (!start_frame.has_value())
 		{
 			delete cb;
 			return nullptr;
@@ -309,10 +219,10 @@ namespace wr::d3d12
 
 		for (std::uint64_t i = 0; i < needed_frames; ++i) 
 		{
-			internal::ClearPage(&(heap->m_bitmap), start_frame + i);
+			util::ClearPage(heap->m_bitmap, start_frame.value() + i);
 		}
 
-		heap->m_current_offset = start_frame * heap->m_alignment;
+		heap->m_current_offset = start_frame.value() * heap->m_alignment;
 
 		cb->m_begin_offset = heap->m_current_offset;
 
@@ -354,67 +264,9 @@ namespace wr::d3d12
 		auto frame_count = heap->m_heap_size / heap->m_alignment;
 		auto needed_frames = aligned_size_in_bytes / heap->m_alignment*heap->m_versioning_count;
 
-		std::uint64_t start_frame = 0;
-		bool counting = false;
-		bool found = false;
-		int free_frames = 0;
+		auto start_frame = util::FindFreePage(heap->m_bitmap, frame_count, needed_frames);
 
-		for (std::uint64_t i = 0; i <= internal::IndexFromBit(frame_count); ++i)
-		{
-			if (heap->m_bitmap[i] != 0Ui64) 
-			{
-				for (std::uint64_t j = 0; j < 64; ++j) 
-				{
-					std::uint64_t to_test = 1Ui64 << j;
-					if (i * 64 + j >= frame_count)
-						break;
-
-					if ((heap->m_bitmap[i] & to_test)) 
-					{
-						if (counting)
-						{
-							free_frames++;
-							if (free_frames == needed_frames)
-							{
-								found = true;
-								break;
-							}
-						}
-						else 
-						{
-							if (needed_frames == 1)
-							{
-								found = true;
-								start_frame = i * 64 + j;
-								break;
-							}
-							else 
-							{
-								start_frame = i * 64 + j;
-								counting = true;
-								free_frames = 1;
-							}
-						}
-					}
-					else 
-					{
-						counting = 0;
-						free_frames = 0;
-					}
-				}
-			}
-			else
-			{
-				counting = false;
-				free_frames = 0;
-			}
-			if (found == true) 
-			{
-				break;
-			}
-		}
-
-		if (found == false) 
+		if (!start_frame.has_value())
 		{
 			delete cb;
 			return nullptr;
@@ -422,10 +274,10 @@ namespace wr::d3d12
 
 		for (std::uint64_t i = 0; i < needed_frames; ++i) 
 		{
-			internal::ClearPage(&(heap->m_bitmap), start_frame + i);
+			util::ClearPage(heap->m_bitmap, start_frame.value() + i);
 		}
 
-		heap->m_current_offset = start_frame * heap->m_alignment;
+		heap->m_current_offset = start_frame.value() * heap->m_alignment;
 
 		CD3DX12_RESOURCE_DESC desc = CD3DX12_RESOURCE_DESC::Buffer(aligned_size_in_bytes, D3D12_RESOURCE_FLAG_NONE);
 		cb->m_gpu_addresses.resize(heap->m_versioning_count);
@@ -480,67 +332,9 @@ namespace wr::d3d12
 		auto frame_count = heap->m_heap_size / heap->m_alignment;
 		auto needed_frames = aligned_size_in_bytes / heap->m_alignment*heap->m_versioning_count;
 
-		std::uint64_t start_frame = 0;
-		bool counting = false;
-		bool found = false;
-		int free_frames = 0;
+		auto start_frame = util::FindFreePage(heap->m_bitmap, frame_count, needed_frames);
 
-		for (std::uint64_t i = 0; i <= internal::IndexFromBit(frame_count); ++i) 
-		{
-			if (heap->m_bitmap[i] != 0Ui64)
-			{
-				for (std::uint64_t j = 0; j < 64; ++j) 
-				{
-					std::uint64_t to_test = 1Ui64 << j;
-					if (i * 64 + j >= frame_count)
-						break;
-
-					if ((heap->m_bitmap[i] & to_test)) 
-					{
-						if (counting) 
-						{
-							free_frames++;
-							if (free_frames == needed_frames)
-							{
-								found = true;
-								break;
-							}
-						}
-						else
-						{
-							if (needed_frames == 1)
-							{
-								found = true;
-								start_frame = i * 64 + j;
-								break;
-							}
-							else 
-							{
-								start_frame = i * 64 + j;
-								counting = true;
-								free_frames = 1;
-							}
-						}
-					}
-					else 
-					{
-						counting = 0;
-						free_frames = 0;
-					}
-				}
-			}
-			else 
-			{
-				counting = false;
-				free_frames = 0;
-			}
-			if (found == true) 
-			{
-				break;
-			}
-		}
-
-		if (found == false) 
+		if (!start_frame.has_value())
 		{
 			delete cb;
 			return nullptr;
@@ -548,10 +342,10 @@ namespace wr::d3d12
 
 		for (std::uint64_t i = 0; i < needed_frames; ++i) 
 		{
-			internal::ClearPage(&(heap->m_bitmap), start_frame + i);
+			util::ClearPage(heap->m_bitmap, start_frame.value() + i);
 		}
 
-		heap->m_current_offset = start_frame * heap->m_alignment;
+		heap->m_current_offset = start_frame.value() * heap->m_alignment;
 				
 		CD3DX12_RESOURCE_DESC desc = CD3DX12_RESOURCE_DESC::Buffer(
 			aligned_size_in_bytes, 
@@ -599,67 +393,9 @@ namespace wr::d3d12
 		auto frame_count = heap->m_heap_size / heap->m_alignment;
 		auto needed_frames = aligned_size_in_bytes / heap->m_alignment*heap->m_versioning_count;
 
-		std::uint64_t start_frame = 0;
-		bool counting = false;
-		bool found = false;
-		int free_frames = 0;
+		auto start_frame = util::FindFreePage(heap->m_bitmap, frame_count, needed_frames);
 
-		for (std::uint64_t i = 0; i <= internal::IndexFromBit(frame_count); ++i)
-		{
-			if (heap->m_bitmap[i] != 0Ui64) 
-			{
-				for (std::uint64_t j = 0; j < 64; ++j) 
-				{
-					std::uint64_t to_test = 1Ui64 << j;
-					if (i * 64 + j >= frame_count)
-						break;
-
-					if ((heap->m_bitmap[i] & to_test)) 
-					{
-						if (counting) 
-						{
-							free_frames++;
-							if (free_frames == needed_frames) 
-							{
-								found = true;
-								break;
-							}
-						}
-						else
-						{
-							if (needed_frames == 1) 
-							{
-								found = true;
-								start_frame = i * 64 + j;
-								break;
-							}
-							else 
-							{
-								start_frame = i * 64 + j;
-								counting = true;
-								free_frames = 1;
-							}
-						}
-					}
-					else 
-					{
-						counting = 0;
-						free_frames = 0;
-					}
-				}
-			}
-			else 
-			{
-				counting = false;
-				free_frames = 0;
-			}
-			if (found == true) 
-			{
-				break;
-			}
-		}
-
-		if (found == false) 
+		if (!start_frame.has_value())
 		{
 			delete cb;
 			return nullptr;
@@ -667,10 +403,10 @@ namespace wr::d3d12
 
 		for (std::uint64_t i = 0; i < needed_frames; ++i) 
 		{
-			internal::ClearPage(&(heap->m_bitmap), start_frame + i);
+			util::ClearPage(heap->m_bitmap, start_frame.value() + i);
 		}
 
-		heap->m_current_offset = start_frame * heap->m_alignment;
+		heap->m_current_offset = start_frame.value() * heap->m_alignment;
 
 		CD3DX12_RESOURCE_DESC desc = CD3DX12_RESOURCE_DESC::Buffer(aligned_size_in_bytes, D3D12_RESOURCE_FLAG_NONE);
 		cb->m_gpu_addresses.resize(heap->m_versioning_count);
@@ -727,7 +463,7 @@ namespace wr::d3d12
 
 		for (int i = 0; i < frame_count; ++i) 
 		{
-			internal::SetPage(&(heap->m_bitmap), frame + i);
+			util::SetPage(heap->m_bitmap, frame + i);
 		}
 
 		delete heapResource;
@@ -766,7 +502,7 @@ namespace wr::d3d12
 
 		for (int i = 0; i < frame_count; ++i) 
 		{
-			internal::SetPage(&(heap->m_bitmap), frame + i);
+			util::SetPage(heap->m_bitmap, frame + i);
 		}
 
 		delete heapResource;
@@ -805,7 +541,7 @@ namespace wr::d3d12
 
 		for (int i = 0; i < frame_count; ++i) 
 		{
-			internal::SetPage(&(heap->m_bitmap), frame + i);
+			util::SetPage(heap->m_bitmap, frame + i);
 		}
 
 		delete heapResource;
